+++ elementapi/ElementApiPlugin.php
@@ -16,7 +16,7 @@ class ElementApiPlugin extends BasePlugin
 	 */
 	public function getVersion()
 	{
-		return '1.3.0';
+		return '1.4.0';
 	}
 
 	/**